+++ ccxt.js
@@ -2675,6 +2675,43 @@ var bitfinex = {
         return this.parseOrder (response);
     },
 
+    getCurrencyName (currency) {
+        if (currency == 'BTC') {
+            return 'bitcoin';
+        } else if (currency == 'LTC') {
+            return 'litecoin';
+        } else if (currency == 'ETH') {
+            return 'ethereum';
+        } else if (currency == 'ETC') {
+            return 'ethereumc';
+        } else if (currency == 'OMNI') {
+            return 'mastercoin'; // ???
+        } else if (currency == 'ZEC') {
+            return 'zcash';
+        } else if (currency == 'XMR') {
+            return 'monero';
+        } else if (currency == 'USD') {
+            return 'wire';
+        } else if (currency == 'DASH') {
+            return 'dash';
+        } else if (currency == 'XRP') {
+            return 'ripple';
+        } else if (currency == 'EOS') {
+            return 'eos';
+        }
+        throw new NotSupported (this.id + ' ' + currency + ' not supported for withdrawal');
+    },
+
+    async withdraw (currency, amount, address, params = {}) {
+        let name = this.getCurrencyName (currency);
+        return this.privatePostWithdraw (this.extend ({
+            'withdraw_type': name,
+            'walletselected': 'exchange',
+            'amount': amount,
+            'address': address,
+        }, params));
+    },
+
     nonce () {
         return this.milliseconds ();
     },